Unifying instrumentation.ts
