+++ e2e/nextjs/src/instrumentation.ts
@@ -1,3 +1,12 @@
-// https://github.com/vercel/next.js/discussions/48273
-// src/instrumentation.ts:
-export { register } from '../instrumentation'
+// instrumentation.ts or src/instrumentation.ts
+import { CONSTANTS } from '@/constants'
+
+export async function register() {
+	const { registerHighlight } = await import('@highlight-run/next/server')
+
+	registerHighlight({
+		projectID: CONSTANTS.NEXT_PUBLIC_HIGHLIGHT_PROJECT_ID,
+		otlpEndpoint: CONSTANTS.NEXT_PUBLIC_HIGHLIGHT_OTLP_ENDPOINT,
+		serviceName: 'my-nextjs-backend',
+	})
+}
